Authentication and authorization

+++ src/stores/auth.ts
@@ -2,25 +2,32 @@ import { ref } from "vue";
 import { defineStore } from "pinia";
 import ApiService from "@/core/services/ApiService";
 import JwtService from "@/core/services/JwtService";
+import AuthService from "@/core/services/AuthService";
 
 export interface User {
   name: string;
   surname: string;
-  email: string;
+  username: string;
   password: string;
-  api_token: string;
+  access_token: string;
+  refresh_token: string;
 }
 
 export const useAuthStore = defineStore("auth", () => {
   const errors = ref({});
-  const user = ref<User>({} as User);
-  const isAuthenticated = ref(!!JwtService.getToken());
+  const user = ref<User>({ name: '', surname: '', username: '', password: '', access_token: '', refresh_token: '' } as User);
+  const isAuthenticated = ref(!!JwtService.getAccessToken());
+
+  function setUsername(username: string) {
+    user.value.username = username;
+  }
 
   function setAuth(authUser: User) {
     isAuthenticated.value = true;
     user.value = authUser;
     errors.value = {};
-    JwtService.saveToken(user.value.api_token);
+    JwtService.saveAccessToken(user.value.access_token);
+    JwtService.saveRefreshToken(user.value.refresh_token);
   }
 
   function setError(error: any) {
@@ -31,12 +38,18 @@ export const useAuthStore = defineStore("auth", () => {
     isAuthenticated.value = false;
     user.value = {} as User;
     errors.value = [];
-    JwtService.destroyToken();
+    JwtService.destroyAccessToken();
+    JwtService.destroyRefreshToken();
+  }
+
+  function decodeToken(token: string) {
+    return JSON.parse(atob(token.split(".")[1]));
   }
 
   function login(credentials: User) {
-    return ApiService.post("/user/login", credentials)
-      .then(({ data }) => {
+    return AuthService.getToken(credentials)
+      .then((data) => {
+        data = { ...data, username: decodeToken(data.access_token).preferred_username };
         setAuth(data);
       })
       .catch(({ response }) => {
@@ -44,8 +57,29 @@ export const useAuthStore = defineStore("auth", () => {
       });
   }
 
+  function refreshToken() {
+    if (JwtService.getRefreshToken()) {
+      AuthService.refreshToken()
+        .then((data) => {
+          data = { ...data, username: decodeToken(data.access_token).preferred_username };
+          setAuth(data);
+        })
+        .catch(({ response }) => {
+          console.log("erro"+response);
+          setError(response.data.errors);
+          purgeAuth();
+        });
+    }
+  }
+
   function logout() {
-    purgeAuth();
+    return AuthService.logout()
+      .then(() => {
+        purgeAuth();
+      })
+      .catch(() => {
+        purgeAuth();
+      });
   }
 
   function register(credentials: User) {
@@ -69,9 +103,9 @@ export const useAuthStore = defineStore("auth", () => {
   }
 
   function verifyAuth() {
-    if (JwtService.getToken()) {
+    if (JwtService.getAccessToken()) {
       ApiService.setHeader();
-      ApiService.post("verify_token", { api_token: JwtService.getToken() })
+      ApiService.post("verify_token", { Access_token: JwtService.getAccessToken() })
         .then(({ data }) => {
           setAuth(data);
         })
@@ -84,6 +118,10 @@ export const useAuthStore = defineStore("auth", () => {
     }
   }
 
+  setInterval(() => {
+    refreshToken();
+  }, 50000);
+
   return {
     errors,
     user,
@@ -92,6 +130,9 @@ export const useAuthStore = defineStore("auth", () => {
     logout,
     register,
     forgotPassword,
+    refreshToken,
     verifyAuth,
+    decodeToken,
+    setUsername
   };
 });
